# src/FullStack.Domain/Models/Response/ProductResponse.cs
namespace FullStack.Domain.Models.Response;

/// <summary>
/// Response model representing a product.
/// </summary>
public record ProductResponse
{
    /// <summary>The product identifier.</summary>
    public int Id { get; init; }

    /// <summary>The product name.</summary>
    public string Name { get; init; } = string.Empty;

    /// <summary>The product price.</summary>
    public decimal Price { get; init; }

    /// <summary>The product description.</summary>
    public string Description { get; init; } = string.Empty;

    /// <summary>The category identifier.</summary>
    public int CategoryId { get; init; }

    /// <summary>The category name.</summary>
    public string CategoryName { get; init; } = string.Empty;

    /// <summary>When the product was created.</summary>
    public DateTimeOffset CreatedAt { get; init; }

    /// <summary>Whether the product is active.</summary>
    public bool IsActive { get; init; }
}